backend/requirements.txt
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-dotenv
psycopg2==2.9.5
PyJWT==2.6.0
requests==2.28.1
django-cors-headers